Revamp of some older tech, expanded Ps_ProcessInformation, added beginning parts for symbol handling via InsightHunter class. Started some work on 1st "telemetry dll' for use sending information back.   Expaned Exports.

+++ FileSandBoxLibrary/InsightSupport.cpp
@@ -76,7 +76,7 @@ InsightSupport_SymbolHandle::~InsightSupport_SymbolHandle()
 bool InsightSupport_SymbolHandle::LoadDll(HANDLE MainDebuggedProcess, LPDEBUG_EVENT EventData)
 {
 	IMAGEHLP_MODULEW64* tmp=0;
-	DWORD Est = 0;
+	DWORD64 Est = 0;
 	tmp = (IMAGEHLP_MODULEW64*)malloc(sizeof(IMAGEHLP_MODULEW64));
 	if (tmp != nullptr)
 	{
